accessibletablesR: Creating accessible tables in an Excel workbook

Creates Excel spreadsheets which align as closely as possible to UK Government Statistical Service's accessibility criteria. Relies upon the R package 'openxlsx'. The package was developed using R 4.1.3, 'dplyr' version 1.1.2, 'openxlsx' version 4.2.5.2, 'conflicted' version 1.2.0, 'stringr' version 1.5.0, 'purrr' version 1.0.1 and 'rlang' version 1.1.0. Alternative options include the R package 'a11ytables' and the Python package 'gptables' (which can be run in R via the package 'reticulate').

See also

Author

Maintainer: Matthew Minifie minifiemj@gmail.com